make sure lotus compiles with +testground build flag

+++ tools/stats/metrics.go
@@ -190,11 +190,11 @@ func RecordTipsetPoints(ctx context.Context, api api.FullNode, pl *PointList, ti
 	}
 	{
 		blks := len(cids)
-		p = NewPoint("chain.gas_fill_ratio", float64(totalGasLimit)/float64(blks*build.BlockGasTarget))
+		p = NewPoint("chain.gas_fill_ratio", float64(totalGasLimit)/float64(int64(blks)*build.BlockGasTarget))
 		pl.AddPoint(p)
-		p = NewPoint("chain.gas_capacity_ratio", float64(totalUniqGasLimit)/float64(blks*build.BlockGasTarget))
+		p = NewPoint("chain.gas_capacity_ratio", float64(totalUniqGasLimit)/float64(int64(blks)*build.BlockGasTarget))
 		pl.AddPoint(p)
-		p = NewPoint("chain.gas_waste_ratio", float64(totalGasLimit-totalUniqGasLimit)/float64(blks*build.BlockGasTarget))
+		p = NewPoint("chain.gas_waste_ratio", float64(totalGasLimit-totalUniqGasLimit)/float64(int64(blks)*build.BlockGasTarget))
 		pl.AddPoint(p)
 	}
